Bedrock/cdk/phoenix/tests/requirements-test.txt
boto3
# arize-phoenix  # Phoenix server
arize-phoenix-otel
openinference-instrumentation-bedrock
opentelemetry-exporter-otlp
